+++ packages/verifier/src/index.js
@@ -1,4 +1,6 @@
+export { derivePortableResultContract } from "./contract.js";
 export { renderExpandedDetails, deriveCompactBanner } from "./display.js";
+export { evaluateFixtureExpectations } from "./expectations.js";
 export { loadFixtureManifest } from "./manifest.js";
 export { resolveVerifierPolicy } from "./policy.js";
 export { verifyFixtureManifest } from "./verify-manifest.js";